==> 1_Scrape_latlon.ipynb <==
# MapQuest API - Latitude and Longitude

With the purpose of having a dataset more complete, and moreover, to be able to display some visualizations in a map, we will extract latitude and longitude for each crime defendant address using [MapQuest API](https://developer.mapquest.com/documentation/)

>If you want to run this code, please create a key in [MapQuest API](https://developer.mapquest.com/documentation/) and add it to **mapquestkey.py** and write in it: 

>`key = "YOUR PASSWORD HERE"` 

Note: Don't forget to add this file to gitignore. 

# Import Dependencies
import requests
import json
from pprint import pprint
import pandas as pd
from mapquestkey import key

# Read the dataset cleaned
csv_path = "Resources/data1_extract.csv"                    
df = pd.read_csv(csv_path)
df.head()

# URL for GET requests to retrieve mapquest data
url = "http://www.mapquestapi.com/geocoding/v1/address?key=IYNehGBC4T9Qt5APqoIcLN7LfJWr7arr&location=5536%20NW%2021ST%20AVENUE%20MIAMI,%20FL%2033142"

# Print the response object to the console
print(requests.get(url))

# Retrieving data and converting it into JSON
print(requests.get(url).json())

# Pretty Print the output of the JSON
response = requests.get(url).json()
# pprint(json.dumps(response, indent=4, sort_keys=True))

# # set up additional columns to hold information
df['lat'] = ""
df['lon'] = ""
df['Zip_code_from_api'] = ""

# df.head()

base_url = "http://open.mapquestapi.com/geocoding/v1/" 

# # use iterrows to iterate through pandas dataframe
for index, row in df.iterrows():

    # get address form df
    address = str(row['Address'])
#     print(address)

    # get city form df
    city = row['City']
#     print(city)
    
    # assemble url and make API request
    # Exmaple URL: http://open.mapquestapi.com/geocoding/v1/address?key=KEY&location=Washington,DC

    url = f"{base_url}address?key={key}&location={address},FL"
#     print(url)
    response = requests.get(url).json()
#   print(response)
    
    # extract results
    results = response['results']
#     pprint(results)

    try:
        
        df.loc[index, 'Zip_code_from_api'] =  results[0]["locations"][0]["postalCode"]
        df.loc[index, 'lat'] = results[0]["locations"][0]["latLng"]['lat']
        df.loc[index, 'lon'] = results[0]["locations"][0]["latLng"]['lng']
        
    except (KeyError, IndexError):
        pass


df.head()

df.to_csv("Resources/data1_extract_clean.csv",index=True,header=True)
