==> 2_Transform_for_ML_part1.ipynb <==
## Transform Dataset for Machine Learning and Visualizations in [Tableau](https://www.tableau.com)

# Import Dependencies
import pandas as pd
import matplotlib.pyplot as plt
from datetime import datetime
%matplotlib inline
import numpy as np

# Import dataset 
df = pd.read_csv('Resources/data1_extract_clean.csv')
df.head()

# Create a new column for Age categories with ranges: 0-25 , 26-50 , >50 where the value is 1 if it is true and 0 if not
df['Age<25'] = np.where(df['Age']<=25, 1,0)
df['Age26-50'] = np.where((df['Age']>25) & (df['Age']<=50), 1,0)
df['Age>50'] = np.where(df['Age']>50, 1,0)

# df.head()

df['Booking_month'] = df['Booking_month'].astype(int)

# Check type of values
df.dtypes

# Create Seasons columns where: Summer= 7, 8, 9. Spring=4,5,6. Winter=1,2,3. Autum=10,11,12.
df['Summer'] = np.where((df['Booking_month']>=7) & (df['Booking_month']<=9), 1,0)
df['Spring'] = np.where((df['Booking_month']>=4) & (df['Booking_month']<=6), 1,0)
df['Winter'] = np.where((df['Booking_month']>=1) & (df['Booking_month']<=3), 1,0)
df['Autum'] = np.where((df['Booking_month']>=10) & (df['Booking_month']<=12), 1,0)

df.tail()

# Drop innecesary columns
df.drop(['Unnamed: 0','Address','City','State','Crime_Family2','Crime_Family','Zip_code_from_api'], axis=1, inplace=True)

df["total_crime"] = df['Age<25'] + df['Age26-50'] + df['Age>50']

df.head()

# Save data to do Viasualizations in Tableau
df.to_csv("Resources/data2_transform_all_crime.csv",index=True,header=True)

### Filter ty Crime Type: ASSAULT & BATTERY

# Filter by State of Florida only (dropping NA also)  
assault = df.loc[df['Crime_Family1'] == "ASSAULT & BATTERY",:]
# df_assault.head()

assault.count()

assault_drop = assault.copy()

# Drop innecesary columns
assault_drop.drop(['Defendant','Zip','Booking_Date','Date_of_birth','Crime_Family1','Booking_month','Age','lat','lon'], axis=1, inplace=True)

# assault_drop.to_csv("Resources/assault_drop.csv",index=True,header=True)

assault_drop.head()

# Create Columns for Days of the week and other categories, in order to apply machine learning model after
assault_dummies = pd.get_dummies(assault_drop)
# df_dummies.head()

# assault_dummies.to_csv("Resources/assault_dummies.csv",index=True,header=True)

assault_combined = assault.copy()

# Create df to prepare data for ML
assault_combined['day_of_week_Monday'] = assault_dummies['day_of_week_Monday']
assault_combined['day_of_week_Tuesday'] = assault_dummies['day_of_week_Tuesday']
assault_combined['day_of_week_Wednesday'] = assault_dummies['day_of_week_Wednesday']
assault_combined['day_of_week_Thursday'] = assault_dummies['day_of_week_Thursday']
assault_combined['day_of_week_Friday'] = assault_dummies['day_of_week_Friday']
assault_combined['day_of_week_Saturday'] = assault_dummies['day_of_week_Saturday']
assault_combined['day_of_week_Sunday'] = assault_dummies['day_of_week_Sunday']
assault_combined['CityRN_FOREIGN'] = assault_dummies['CityRN_FOREIGN']
assault_combined['CityRN_HOMELESS'] = assault_dummies['CityRN_HOMELESS']
assault_combined['CityRN_MIAMI'] = assault_dummies['CityRN_MIAMI']

assault_combined.head()

# assault_combined.to_csv("Resources/assault_combined.csv",index=True,header=True)

# Drop innecesary columns
assault_combined.drop(['Unnamed: 0.1','Defendant','Zip','Booking_Date','Date_of_birth','Age','day_of_week','Crime_Family1','CityRN'], axis=1, inplace=True)

assault_group = assault_combined.groupby('Booking_year_month').sum()

assault_group.head()

assault_group.reset_index(inplace=True)

# assault_group['Booking'] = assault_group['Booking_year_month']

assault_group.head()

assault_group.to_csv("Resources/data2_transform_assault_group1.csv",index=True,header=True)

### Merge with Unemployment rate only will work if we do the ML with all crimes. 

# df_unem = pd.read_csv('Resources/Florida_unemployment_rate.csv')
# df_unem.head()

# df_unem.rename(columns={"FLMIAM6URN":"Unemploymnet Rate"}, inplace = True)

# df_merge_assault = pd.merge(df_group, df_unem, left_on="Booking_year_month",right_on="DATE")
# df_merge_assault.head()

# # Drop innecesary columns
# df_merge_assault.drop(['DATE'], axis=1, inplace=True)

# df_merge_assault.to_csv("Resources/data2_transform_assault_group2.csv",index=True,header=True)
