==> 2_Transform_for_ML_part2.ipynb <==
## Transform for Machine Learning 

import pandas as pd
# import matplotlib.pylab as plt
import matplotlib.pyplot as plt
from datetime import datetime
%matplotlib inline
import numpy as np

df = pd.read_csv('Resources/data2_transform_assault_group1.csv')
df.head()

# Add column of integers based on Booking_year_month to convert the table after
df['month_index'] = np.arange(len(df))+1

# df.head()

df['Booking_year_month_original'] = df['Booking_year_month'] 

df.columns

i = 1
for index, row in df["month_index"].iteritems():
    if int(row) == i:
        df.loc[index, 'total_crime_3'] = df.iloc[i-4]['total_crime'] #3 months earlier
    if int(row) == i:
        df.loc[index, 'age<25_3'] = df.iloc[i-4]['Age<25']
    if int(row) == i:
        df.loc[index, 'age26-50_3'] = df.iloc[i-4]['Age26-50']
    if int(row) == i:
        df.loc[index, 'age>50_3'] = df.iloc[i-4]['Age>50']
    if int(row) == i:
        df.loc[index, 'day_of_week_Monday_3'] = df.iloc[i-4]['day_of_week_Monday']
    if int(row) == i:
        df.loc[index, 'day_of_week_Tuesday_3'] = df.iloc[i-4]['day_of_week_Tuesday']
    if int(row) == i:
        df.loc[index, 'day_of_week_Wednesday_3'] = df.iloc[i-4]['day_of_week_Wednesday']
    if int(row) == i:
        df.loc[index, 'day_of_week_Thursday_3'] = df.iloc[i-4]['day_of_week_Thursday']
    if int(row) == i:
        df.loc[index, 'day_of_week_Friday_3'] = df.iloc[i-4]['day_of_week_Friday']
    if int(row) == i:
        df.loc[index, 'day_of_week_Saturday_3'] = df.iloc[i-4]['day_of_week_Saturday']
    if int(row) == i:
        df.loc[index, 'day_of_week_Sunday_3'] = df.iloc[i-4]['day_of_week_Sunday']
    if int(row) == i:
        df.loc[index, 'summer_3'] = df.iloc[i-4]['Summer']
    if int(row) == i:
        df.loc[index, 'spring_3'] = df.iloc[i-4]['Spring']
    if int(row) == i:
        df.loc[index, 'winter_3'] = df.iloc[i-4]['Winter']
    if int(row) == i:
        df.loc[index, 'autum_3'] = df.iloc[i-4]['Autum']
    if int(row) == i:
        df.loc[index, 'CityRN_FOREIGN_3'] = df.iloc[i-4]['CityRN_FOREIGN']
    if int(row) == i:
        df.loc[index, 'CityRN_HOMELESS_3'] = df.iloc[i-4]['CityRN_HOMELESS']
    if int(row) == i:
        df.loc[index, 'CityRN_MIAMI_3'] = df.iloc[i-4]['CityRN_MIAMI']
    if int(row) == i:
        df.loc[index, 'total_crime_2'] = df.iloc[i-3]['total_crime'] #2 months earlier
    if int(row) == i:
        df.loc[index, 'Age<25_2'] = df.iloc[i-3]['Age<25'] 
    if int(row) == i:
        df.loc[index, 'Age26-50_2'] = df.iloc[i-3]['Age26-50']
    if int(row) == i:
        df.loc[index, 'Age>50_2'] = df.iloc[i-3]['Age>50']
    if int(row) == i:
        df.loc[index, 'day_of_week_Monday_2'] = df.iloc[i-3]['day_of_week_Monday']
    if int(row) == i:
        df.loc[index, 'day_of_week_Tuesday_2'] = df.iloc[i-3]['day_of_week_Tuesday']
    if int(row) == i:
        df.loc[index, 'day_of_week_Wednesday_2'] = df.iloc[i-3]['day_of_week_Wednesday']
    if int(row) == i:
        df.loc[index, 'day_of_week_Thursday_2'] = df.iloc[i-3]['day_of_week_Thursday']
    if int(row) == i:
        df.loc[index, 'day_of_week_Friday_2'] = df.iloc[i-3]['day_of_week_Friday']
    if int(row) == i:
        df.loc[index, 'day_of_week_Saturday_2'] = df.iloc[i-3]['day_of_week_Saturday']
    if int(row) == i:
        df.loc[index, 'day_of_week_Sunday_2'] = df.iloc[i-3]['day_of_week_Sunday']
    if int(row) == i:
        df.loc[index, 'summer_2'] = df.iloc[i-3]['Summer']
    if int(row) == i:
        df.loc[index, 'spring_2'] = df.iloc[i-3]['Spring']
    if int(row) == i:
        df.loc[index, 'winter_2'] = df.iloc[i-3]['Winter']
    if int(row) == i:
        df.loc[index, 'autum_2'] = df.iloc[i-3]['Autum']
    if int(row) == i:
        df.loc[index, 'CityRN_FOREIGN_2'] = df.iloc[i-3]['CityRN_FOREIGN']
    if int(row) == i:
        df.loc[index, 'CityRN_HOMELESS_2'] = df.iloc[i-3]['CityRN_HOMELESS']
    if int(row) == i:
        df.loc[index, 'CityRN_MIAMI_2'] = df.iloc[i-3]['CityRN_MIAMI']
    if int(row) == i:
        df.loc[index, 'total_crime_1'] = df.iloc[i-2]['total_crime']   #1 months earlier
    if int(row) == i:
        df.loc[index, 'age<25_1'] = df.iloc[i-2]['Age<25']
    if int(row) == i:
        df.loc[index, 'age26-50_1'] = df.iloc[i-2]['Age26-50']
    if int(row) == i:
        df.loc[index, 'age>50_1'] = df.iloc[i-2]['Age>50']
    if int(row) == i:
        df.loc[index, 'day_of_week_Monday_1'] = df.iloc[i-2]['day_of_week_Monday']
    if int(row) == i:
        df.loc[index, 'day_of_week_Tuesday_1'] = df.iloc[i-2]['day_of_week_Tuesday']
    if int(row) == i:
        df.loc[index, 'day_of_week_Wednesday_1'] = df.iloc[i-2]['day_of_week_Wednesday']
    if int(row) == i:
        df.loc[index, 'day_of_week_Thursday_1'] = df.iloc[i-2]['day_of_week_Thursday']
    if int(row) == i:
        df.loc[index, 'day_of_week_Friday_1'] = df.iloc[i-2]['day_of_week_Friday']
    if int(row) == i:
        df.loc[index, 'day_of_week_Saturday_1'] = df.iloc[i-2]['day_of_week_Saturday']
    if int(row) == i:
        df.loc[index, 'day_of_week_Sunday_1'] = df.iloc[i-2]['day_of_week_Sunday']
    if int(row) == i:
        df.loc[index, 'summer_1'] = df.iloc[i-2]['Summer']
    if int(row) == i:
        df.loc[index, 'spring_1'] = df.iloc[i-2]['Spring']
    if int(row) == i:
        df.loc[index, 'winter_1'] = df.iloc[i-2]['Winter']
    if int(row) == i:
        df.loc[index, 'autum_1'] = df.iloc[i-2]['Autum']
    if int(row) == i:
        df.loc[index, 'CityRN_FOREIGN_1'] = df.iloc[i-2]['CityRN_FOREIGN']
    if int(row) == i:
        df.loc[index, 'CityRN_HOMELESS_1'] = df.iloc[i-2]['CityRN_HOMELESS']
    if int(row) == i:
        df.loc[index, 'CityRN_MIAMI_1'] = df.iloc[i-2]['CityRN_MIAMI']
        
        
        i = i + 1

df.head()

# df.to_csv("Resources/df_ml_1_categories.csv",index=True,header=True)

# Drop innecesary columns
df.drop(['Age<25','Age26-50','Age>50','Booking_month','Summer','Spring','Winter','Autum','day_of_week_Monday','day_of_week_Tuesday','day_of_week_Wednesday','day_of_week_Thursday',\
        'day_of_week_Friday','day_of_week_Saturday','day_of_week_Sunday','CityRN_FOREIGN','CityRN_HOMELESS','CityRN_MIAMI','month_index','Booking_year_month_original',\
        'lat','lon'], axis=1, inplace=True)

# df.head()

df.drop(df.index[[0,1,2]], inplace=True)

df.head()

# Save final dataset ready to apply Machine Learning
df.to_csv("Resources/data3_load_assault_group_to_ML.csv",index=True,header=True)
