==> 2_Transform_for_ML_part2_rolling_one_month_behind.ipynb <==
## Transform for Machine Learning 

import pandas as pd
# import matplotlib.pylab as plt
import matplotlib.pyplot as plt
from datetime import datetime
%matplotlib inline
import numpy as np

df = pd.read_csv('Resources/data2_transform_assault_group1.csv')
df.head()

# Add column of integers based on Booking_year_month to convert the table after
df['month_index'] = np.arange(len(df))+1

# df.head()

df['Booking_year_month_original'] = df['Booking_year_month'] 

df.columns

i = 1
for index, row in df["month_index"].iteritems():
    if int(row) == i:
        df.loc[index, 'total_crime_1'] = df.iloc[i-2]['total_crime']   #1 months earlier
    if int(row) == i:
        df.loc[index, 'age<25_1'] = df.iloc[i-2]['Age<25']
    if int(row) == i:
        df.loc[index, 'age26-50_1'] = df.iloc[i-2]['Age26-50']
    if int(row) == i:
        df.loc[index, 'age>50_1'] = df.iloc[i-2]['Age>50']
    if int(row) == i:
        df.loc[index, 'day_of_week_Monday_1'] = df.iloc[i-2]['day_of_week_Monday']
    if int(row) == i:
        df.loc[index, 'day_of_week_Tuesday_1'] = df.iloc[i-2]['day_of_week_Tuesday']
    if int(row) == i:
        df.loc[index, 'day_of_week_Wednesday_1'] = df.iloc[i-2]['day_of_week_Wednesday']
    if int(row) == i:
        df.loc[index, 'day_of_week_Thursday_1'] = df.iloc[i-2]['day_of_week_Thursday']
    if int(row) == i:
        df.loc[index, 'day_of_week_Friday_1'] = df.iloc[i-2]['day_of_week_Friday']
    if int(row) == i:
        df.loc[index, 'day_of_week_Saturday_1'] = df.iloc[i-2]['day_of_week_Saturday']
    if int(row) == i:
        df.loc[index, 'day_of_week_Sunday_1'] = df.iloc[i-2]['day_of_week_Sunday']
    if int(row) == i:
        df.loc[index, 'summer_1'] = df.iloc[i-2]['Summer']
    if int(row) == i:
        df.loc[index, 'spring_1'] = df.iloc[i-2]['Spring']
    if int(row) == i:
        df.loc[index, 'winter_1'] = df.iloc[i-2]['Winter']
    if int(row) == i:
        df.loc[index, 'autum_1'] = df.iloc[i-2]['Autum']
    if int(row) == i:
        df.loc[index, 'CityRN_FOREIGN_1'] = df.iloc[i-2]['CityRN_FOREIGN']
    if int(row) == i:
        df.loc[index, 'CityRN_HOMELESS_1'] = df.iloc[i-2]['CityRN_HOMELESS']
    if int(row) == i:
        df.loc[index, 'CityRN_MIAMI_1'] = df.iloc[i-2]['CityRN_MIAMI']
        
        
        i = i + 1

df.head()

# df.to_csv("Resources/df_ml_1_categories.csv",index=True,header=True)

# Drop innecesary columns
df.drop(['Unnamed: 0','Age<25','Age26-50','Age>50','Booking_month','Summer','Spring','Winter','Autum','day_of_week_Monday','day_of_week_Tuesday','day_of_week_Wednesday','day_of_week_Thursday',\
        'day_of_week_Friday','day_of_week_Saturday','day_of_week_Sunday','CityRN_FOREIGN','CityRN_HOMELESS','CityRN_MIAMI','month_index','Booking_year_month_original',\
        'lat','lon'], axis=1, inplace=True)

# df.head()

df.drop(df.index[[0,1,2]], inplace=True)

df.head()

# Change seasons values to determine wchich months belongs to each season 
for index, row in df["summer_1"].iteritems():
    if int(row) >0:
        df.loc[index, 'summer_1'] = 1 
    else:
        df.loc[index, 'summer_1'] = 0

for index, row in df["spring_1"].iteritems():
    if int(row) >0:
        df.loc[index, 'spring_1'] = 1 
    else:
        df.loc[index, 'spring_1'] = 0
        
for index, row in df["winter_1"].iteritems():
    if int(row) >0:
        df.loc[index, 'winter_1'] = 1 
    else:
        df.loc[index, 'winter_1'] = 0

for index, row in df["autum_1"].iteritems():
    if int(row) >0:
        df.loc[index, 'autum_1'] = 1 
    else:
        df.loc[index, 'autum_1'] = 0
        

for index, row in df["summer_2"].iteritems():
    if int(row) >0:
        df.loc[index, 'summer_2'] = 1 
    else:
        df.loc[index, 'summer_2'] = 0

for index, row in df["spring_2"].iteritems():
    if int(row) >0:
        df.loc[index, 'spring_2'] = 1 
    else:
        df.loc[index, 'spring_2'] = 0
        
for index, row in df["winter_2"].iteritems():
    if int(row) >0:
        df.loc[index, 'winter_2'] = 1 
    else:
        df.loc[index, 'winter_2'] = 0

for index, row in df["autum_2"].iteritems():
    if int(row) >0:
        df.loc[index, 'autum_2'] = 1 
    else:
        df.loc[index, 'autum_2'] = 0
        

for index, row in df["summer_3"].iteritems():
    if int(row) >0:
        df.loc[index, 'summer_3'] = 1 
    else:
        df.loc[index, 'summer_3'] = 0

for index, row in df["spring_3"].iteritems():
    if int(row) >0:
        df.loc[index, 'spring_3'] = 1 
    else:
        df.loc[index, 'spring_3'] = 0
        
for index, row in df["winter_3"].iteritems():
    if int(row) >0:
        df.loc[index, 'winter_3'] = 1 
    else:
        df.loc[index, 'winter_3'] = 0

for index, row in df["autum_3"].iteritems():
    if int(row) >0:
        df.loc[index, 'autum_3'] = 1 
    else:
        df.loc[index, 'autum_3'] = 0

# Drop innecesary columns
df.drop(['DOB_month'], axis=1, inplace=True)

df.head()

# Save final dataset ready to apply Machine Learning
df.to_csv("Resources/data3_load_assault_group_to_ML_one_month_behind.csv",index=True,header=True)
