==> 3_Load_to_ML_test2_caro2.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
%matplotlib inline
import numpy as np

# Data base past month
df = pd.read_csv('Resources/data3_test.csv',parse_dates =["Booking_year_month"])
df.head()

# Drop innecesary columns
df.drop(['Unnamed: 0'], axis=1, inplace=True)
df['total_crime']= df['total_crime'].astype(float)
